fundamentals of measurement controlling
