+++ src/services/measurement_service.py
@@ -1,0 +1,7 @@
+from socketio import AsyncServer
+
+from sio.models import Lobby
+
+
+def measurement_controller(sio: AsyncServer, lobby: Lobby):
+    pass
