+++ src/sio/models.py
@@ -1,0 +1,21 @@
+from pydantic import BaseModel
+from typing import List, Dict
+
+
+class LobbyClient(BaseModel):
+    sid: str
+    index: int
+
+
+class Lobby(BaseModel):
+    host: str
+    lobby_id: str
+    microphones: List[LobbyClient]
+    speakers: List[LobbyClient]
+
+
+class SocketSession(BaseModel):
+    lobby: str
+    isHost: bool
+
+lobbies: Dict[str, Lobby] = {}
